+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 4.1 - Unreleased
 ----------------
 
+- Allow to skip setuptools extensions if a MANIFEST.in template exists.
+  [stefan]
 
 4.0.1 - 2017-07-20
 ------------------