+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog
 - Add -m option to skip setuptools extensions.
   [stefan]
 
+- Never read existing ``SOURCES.txt`` files.
+  [stefan]
+
 4.0.1 - 2017-07-20
 ------------------